Move ShareMode and ReceiveMode into Mode module
