+++ onionshare_gui/mode/receive_mode/uploads.py
@@ -24,7 +24,7 @@ from datetime import datetime
 from PyQt5 import QtCore, QtWidgets, QtGui
 
 from onionshare import strings
-from ..widgets import Alert
+from ...widgets import Alert
 
 
 class File(QtWidgets.QWidget):
