+++ onionshare_gui/mode/share_mode/info.py
